app/requirements.txt
streamlit==1.23.1
pymilvus==2.2.10
tika==1.24
sentence-transformers==2.2.2
langchain==0.0.208
huggingface_hub==0.15.1
faiss-cpu==1.7.3
python-dotenv==1.0.0
